requirements.txt
streamlit_cookies_controller
streamlit_option_menu
streamlit-toggle-switch
streamlit-aggrid